move field models into separate files
